LayoutTests/platform/ios-simulator/webaudio/analyser-exception-expected.txt

Tests that AnalyserNode validates minDecibels, maxDecibels and smoothingTimeConstant values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS analyser.minDecibels = -20 threw exception Error: IndexSizeError: DOM Exception 1.
PASS analyser.maxDecibels = -120 threw exception Error: IndexSizeError: DOM Exception 1.
PASS analyser.smoothingTimeConstant = 2 threw exception Error: IndexSizeError: DOM Exception 1.
PASS successfullyParsed is true

TEST COMPLETE
